refactor: reduced cognitive complexity and convert hook to ts
